requirements.txt
#spatial tracker v2
numpy==1.26.4
gradio==5.31.0
pydantic==2.10.6
opencv-python-headless==4.10.0.84
einops
einx
plotly
prettytable
decord
easydict
flow_vis
moviepy==1.0.0
safetensors
scikit-learn
hydra-core
omegaconf
albumentations
matplotlib
mediapy
scikit-image
pycolmap==3.11.1
git+https://github.com/facebookresearch/segment-anything.git
git+https://github.com/EasternJournalist/utils3d.git#egg=utils3d
huggingface_hub
pyceres
kornia==0.8.1
xformers==0.0.28
timm
PyJWT
gdown
rich
decord
ray[default]
jaxtyping
transformers

#trellis
spaces
trimesh==4.5.3
pytorch_lightning
open3d
deepspeed
git+https://github.com/NVlabs/nvdiffrast
xatlas
igraph
pyvista
pymeshfix
peft
spconv-cu121==2.3.7

#foundationpose
h5py==3.10.0
transformations==2024.6.1
ruamel.yaml==0.18.6
warp-lang==1.0.2

#amodal3r
rembg
onnxruntime
